requirements.txt
coverage
flake8
Flask
Flask-RESTful
Flask-Bcrypt
Flask-Bootstrap
Flask-DebugToolbar
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-Testing
Flask-WTF
requests
psycopg2
